+++ src/Actions/ViewAction.php
@@ -19,14 +19,14 @@ class ViewAction extends BaseViewAction
             fn (FullCalendarWidget $livewire) => $livewire->getRecord()
         );
 
-        $this->form(
+        $this->schema(
             fn (FullCalendarWidget $livewire) => $livewire->getFormSchema()
         );
 
         $this->modalFooterActions(
             fn (ViewAction $action, FullCalendarWidget $livewire) => [
-                ...$livewire->getCachedModalActions(),
-//                $action->getModalCancelAction(),
+                ...$livewire->getCachedFormActions(),
+                $action->getModalCancelAction(),
             ]
         );