+++ src/Widgets/Concerns/IsBackwardCompatible.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SchultenMedia\FilamentFullCalendar\Widgets\Concerns;
+
+trait IsBackwardCompatible
+{
+    protected function getHeaderActions(): array
+    {
+        return $this->headerActions();
+    }
+
+    protected function getFormActions(): array
+    {
+        return $this->modalActions();
+    }
+}