+++ src/Widgets/FullCalendarWidget.php
@@ -7,6 +7,8 @@ use Filament\Actions\Concerns\InteractsWithActions;
 use Filament\Actions\Contracts\HasActions;
 use Filament\Forms\Concerns\InteractsWithForms;
 use Filament\Forms\Contracts\HasForms;
+use Filament\Pages\Concerns\InteractsWithFormActions;
+use Filament\Pages\Concerns\InteractsWithHeaderActions;
 use Filament\Widgets\Widget;
 use SchultenMedia\FilamentFullCalendar\Actions;
 
@@ -16,12 +18,16 @@ class FullCalendarWidget extends Widget implements HasForms, HasActions
     use InteractsWithActions;
     use Concerns\InteractsWithEvents;
     use Concerns\InteractsWithRecords;
-    use Concerns\InteractsWithHeaderActions;
-    use Concerns\InteractsWithModalActions;
+    use InteractsWithHeaderActions;
+    use InteractsWithFormActions;
     use Concerns\InteractsWithRawJS;
     use Concerns\CanBeConfigured;
+    use Concerns\IsBackwardCompatible{
+        Concerns\IsBackwardCompatible::getHeaderActions insteadof InteractsWithHeaderActions;
+        Concerns\IsBackwardCompatible::getFormActions insteadof InteractsWithFormActions;
+    }
 
-    protected static string $view = 'filament-fullcalendar::fullcalendar';
+    protected string $view = 'filament-fullcalendar::fullcalendar';
 
     protected int | string | array $columnSpan = 'full';