+++ src/helpers.php
@@ -2,7 +2,7 @@
 
 namespace SchultenMedia\FilamentFullCalendar;
 
-if(! function_exists('SchultenMedia\FilamentFullCalendar\array_merge_recursive_unique')) {
+if (! function_exists('SchultenMedia\FilamentFullCalendar\array_merge_recursive_unique')) {
     function array_merge_recursive_unique(array $array1, array $array2): array
     {
         foreach ($array2 as $key => $value) {